+++ yellowbox/extras/redis.py
@@ -1,4 +1,3 @@
-from contextlib import contextmanager
 from typing import TypeVar, Callable, Mapping, Union, Sequence, IO, Optional
 
 from docker import DockerClient
@@ -10,7 +9,6 @@ from yellowbox.retry import RetrySpec
 from yellowbox.subclasses import SingleContainerService
 
 __all__ = ['RedisService', 'REDIS_DEFAULT_PORT', 'DEFAULT_RDB_PATH', 'append_state']
-
 
 REDIS_DEFAULT_PORT = 6379
 DEFAULT_RDB_PATH = "/data/dump.rdb"
@@ -60,18 +58,9 @@ class RedisService(SingleContainerService, RunMixin):
         self.started = True
         return self
 
-    @contextmanager
-    def clean_slate(self):
-        """
-        Notes:
-            This feature is experimental.
-        """
+    def reset_state(self):
         client: Redis
         with self.client() as client:
-            keys = client.keys()
-            if keys:
-                raise RuntimeError(f'Redis db is not empty (found keys {keys})')
-            yield
             client.flushall()
 
     def set_state(self, db_dict: RedisState):